requirements.txt
customtkinter
pycryptodome
pycryptodomex
tkinterdnd2
pyinstaller
